Updates to files per Johnathan's comments
